Move interface from track to cmd pkg
The interface is only used within the command package.
