+++ cmd/cmd.go
@@ -1,5 +1,6 @@
-package track
+package cmd
 
+// ConfigSerializer reads and serializes JSON configs.
 type ConfigSerializer interface {
 	NewConfigFromFile(string) error
 	ToJSON() ([]byte, error)
